Support fuer Hints

+++ src/map.c
@@ -157,7 +157,7 @@ void explore_area(Spawn* spawn, Map* map) {
 	
 	for(i = y; i < y + VISUAL_SQUARE; ++i) {
 		for(j = x; j < x + VISUAL_SQUARE; ++j) {
-			if(x >= 0 && y >= 0 && x < map->x && y < map->y) {
+			if(x >= 0 && y >= 0 && (unsigned int)x < map->x && (unsigned int)y < map->y) {
 				map->tiles[i * map->x + j].spotted = 1;
 			}
 		}
